qr_backend/requirements.txt
Django>=5.0
djangorestframework
qrcode
Pillow
django-cors-headers